Move some cache-related code to a separate file
