+++ src/breakfast.py
@@ -17,6 +17,8 @@ from networkx.algorithms.components.connected import connected_components
 from scipy.sparse import csr_matrix
 from sklearn.metrics import pairwise_distances_chunked
 
+import cache as ca
+
 
 # Merge connected components
 def _to_graph(l):
@@ -136,64 +138,10 @@ def calc_sparse_matrix(meta, args):
             print("Import from pickle file")
             cache = cPickle.load(f)
 
-        max_dist_cached = cache["max_dist"]
+        ca.validate(cache, args, VERSION)
 
-        if args.max_dist != max_dist_cached:
-            print(
-                f"WARNING: Cached results were created using a differnt max-dist paramter"
-            )
-            print(
-                f"Current max-dist parameter: {args.max_dist} \n Cached max-dist parameter: {max_dist_cached}"
-            )
-            raise UnboundLocalError()
-
-        version_cached = cache["version"]
-        if version_cached != VERSION:
-            print(
-                f"WARNING: Cached results were created using breakfast version {version_cached}"
-            )
-
-        # compare cached IDs and current IDs to check if sequences have been deleted
-        meta_cached = cache["meta"]
-        cached_seqs = meta_cached["feature"]
-        cached_id = meta_cached["id"]
-        current_id = meta["id"].tolist()
-
-        # flatten list of IDs to compare cached and current IDs
-        # since we summarized IDs with identical sequences to sets
-        # e.g. [(ID1, ID2), (ID3)] -> {ID1, ID2, ID3}
-        flat_cached_id = set(item for sublist in cached_id for item in sublist)
-        flat_current_id = set(item for sublist in current_id for item in sublist)
-
-        # get all IDs which are part of cached ID but not anymore of current ID
-        del_seqs = list(flat_cached_id.difference(flat_current_id))
-        new_seqs = list(flat_current_id.difference(flat_cached_id))
-        common_seqs = list(flat_cached_id.intersection(flat_current_id))
-
-        # Undo the grouping/deduplication of sequences so we can work with
-        # individual sequence ids
-        def ungroup_df(df):
-            df_ungrouped = pd.DataFrame(
-                {
-                    "id": np.concatenate(df["id"].values),
-                    "feature": np.repeat(df["feature"].values, df["id"].str.len()),
-                }
-            )
-            return df_ungrouped
-
-        meta_cached_ungrouped = ungroup_df(meta_cached).set_index("id")
-        meta_new_ungrouped = ungroup_df(meta).set_index("id")
-
-        meta_merged = meta_cached_ungrouped.join(
-            meta_new_ungrouped, how="inner", lsuffix="_cached", rsuffix="_new"
-        )
-        meta_merged["modified"] = (
-            meta_merged["feature_cached"] != meta_merged["feature_new"]
-        )
-        mod_seqs = meta_merged[meta_merged["modified"]].index.tolist()
-
-        print(f"{len(mod_seqs)} modified sequence(s)")
-        print(f"The following sequences were modified {mod_seqs}")
+        del_seqs, new_seqs, common_seqs = ca.find_deleted(cache, meta)
+        mod_seqs = ca.find_modified(cache, meta)
 
         # Update the deleted seqs list with the sequences that were
         # modified, and also add them to the "new_seqs" list so that they
@@ -213,6 +161,7 @@ def calc_sparse_matrix(meta, args):
         new_seqs_grouped = []
         new_seqs_grouped_idx = []
         del_seqs_grouped_idx = []
+        current_id = meta["id"].tolist()
         # go trough all current ID sets
         for grouped_seqs_idx, grouped_seqs in enumerate(current_id):
             counter_new_seq = 0
@@ -225,6 +174,8 @@ def calc_sparse_matrix(meta, args):
                 new_seqs_grouped.append(grouped_seqs)
                 new_seqs_grouped_idx.append(grouped_seqs_idx)
 
+        meta_cached = cache["meta"]
+        cached_id = meta_cached["id"]
         if (len(del_seqs)) > 0:
             print(f"{len(del_seqs)} deleted sequence(s)")
             print(f"The following sequences got deleted {del_seqs}")
